+++ synapse-core/src/main/java/de/otto/synapse/messagestore/MessageStoreFactory.java
@@ -11,8 +11,8 @@ public interface MessageStoreFactory<T extends MessageStore> {
     /**
      * Creates and returns a {@link MessageStore}.
      *
-     * @param channelName the name of the channel associated to the MessageStore.
-     *
+     * @param name the name of the {@code MessageStore}
+     * @param channelName the name of the channel associated to the {@code MessageStore}.
      * @return MessageStore
      */
     T createMessageStoreFor(@Nonnull String name, @Nonnull String channelName);